+++ src/Amazon.S3/Model/S3MultipartUploadOperation_Internal.m
@@ -254,7 +254,7 @@ typedef void (^AbortMultipartUploadBlock)();
         }
         else if([response isKindOfClass:[S3UploadPartResponse class]])
         {
-            AMZLogDebug(@"UploadPart succeeded: %d", self.currentPartNo);
+            AMZLogDebug(@"UploadPart succeeded: %ld", (unsigned long)self.currentPartNo);
 
             S3UploadPartResponse *uploadPartResponse = (S3UploadPartResponse *)self.response;
